refactor: extract SettingsGitHubExportPane.swift
Extracts the SettingsGitHubExportPane View (~205 lines) from
SettingsIntegrationsPanes.swift into its own file. SettingsJiraExportPane
remains in SettingsIntegrationsPanes.swift for now (extraction stacked
separately in #685).

Byte-identical move of the pane; new SettingsIntegrationsPanes.swift
retains only the Jira pane.

SettingsIntegrationsPanes.swift: 484 → 279 lines. Tests: 667.

⚠️ Manual smoke check recommended: render Settings > GitHub Export
pane and verify layout / behavior unchanged.
